+++ behave/model.py
@@ -187,11 +187,34 @@ class BasicStatement(object):
 
 
 class TagStatement(BasicStatement):
+
     def __init__(self, filename, line, keyword, name, tags):
         super(TagStatement, self).__init__(filename, line, keyword, name)
-        if tags is None:
-            tags = []
         self.tags = tags
+
+
+class TagAndStatusStatement(BasicStatement):
+    final_status = ('passed', 'failed', 'skipped')
+
+    def __init__(self, filename, line, keyword, name, tags):
+        super(TagAndStatusStatement, self).__init__(filename, line, keyword, name)
+        self.tags = tags
+        self.should_skip = False
+        self._cached_status = None
+
+    @property
+    def status(self):
+        if self._cached_status not in self.final_status:
+            # -- RECOMPUTE: As long as final status is not reached.
+            self._cached_status = self.compute_status()
+        return self._cached_status
+
+    def reset(self):
+        self.should_skip = False
+        self._cached_status = None
+
+    def compute_status(self):
+        raise NotImplementedError
 
 
 class Replayable(object):
@@ -201,7 +224,7 @@ class Replayable(object):
         getattr(formatter, self.type)(self)
 
 
-class Feature(TagStatement, Replayable):
+class Feature(TagAndStatusStatement, Replayable):
     '''A `feature`_ parsed from a *feature file*.
 
     The attributes are:
@@ -268,20 +291,25 @@ class Feature(TagStatement, Replayable):
 
     type = "feature"
 
-    # XXX-JE-ORIG: def __init__(self, filename, line, keyword, name, tags=[], description=[],
-    # XXX-JE-ORIG:              scenarios=[], background=None):
-    def __init__(self, filename, line, keyword, name, tags=None, description=None,
-                 scenarios=None, background=None):
+    def __init__(self, filename, line, keyword, name, tags=None,
+                 description=None, scenarios=None, background=None):
+        tags = tags or []
         super(Feature, self).__init__(filename, line, keyword, name, tags)
         self.description = description or []
         self.scenarios = []
         self.background = background
         self.parser = None
-        if scenarios is None:
-            scenarios = []
+        if scenarios:
+            for scenario in scenarios:
+                self.add_scenario(scenario)
 
-        for scenario in scenarios:
-            self.add_scenario(scenario)
+    def reset(self):
+        '''
+        Reset to clean state before a test run.
+        '''
+        super(Feature, self).reset()
+        for scenario in self.scenarios:
+            scenario.reset()
 
     def __repr__(self):
         return '<Feature "%s": %d scenario(s)>' % \
@@ -295,48 +323,37 @@ class Feature(TagStatement, Replayable):
         scenario.background = self.background
         self.scenarios.append(scenario)
 
-    @property
-    def status(self):
+    def compute_status(self):
+        """
+        Compute the status of this feature based on its:
+           * scenarios
+           * scenario outlines
+
+        :return: Computed status (as string-enum).
+        """
         skipped = True
         passed_count = 0
-        for scenario_or_outline in self.scenarios:
-            # FIXME: Check if necessary, ScenarioOutline.status computes OK.
-            if isinstance(scenario_or_outline, ScenarioOutline):
-                for scenario in scenario_or_outline:
-                    if scenario.status == 'failed':
-                        return 'failed'
-                    if scenario.status == 'untested':
-                        if passed_count > 0:
-                            return 'failed'  # ABORTED: Some passed, ...
-                        return 'untested'
-                    if scenario.status != 'skipped':
-                        skipped = False
-            else:
-                scenario = scenario_or_outline
-                if scenario.status == 'failed':
-                    return 'failed'
-                if scenario.status == 'untested':
-                    if passed_count > 0:
-                        return 'failed'  # ABORTED: Some passed, now untested.
-                    return 'untested'
-                if scenario.status != 'skipped':
-                    skipped = False
-            if scenario_or_outline.status == 'passed':
+        for scenario in self.scenarios:
+            scenario_status = scenario.status
+            if scenario_status == 'failed':
+                return 'failed'
+            elif scenario_status == 'untested':
+                if passed_count > 0:
+                    return 'failed'  # ABORTED: Some passed, now untested.
+                return 'untested'
+            if scenario_status != 'skipped':
+                skipped = False
+            if scenario_status == 'passed':
                 passed_count += 1
         return skipped and 'skipped' or 'passed'
 
     @property
     def duration(self):
-        # -- ORIG:
-        # if self.background:
-        #     duration = self.background.duration or 0.0
-        # else:
-        #     duration = 0.0
         # -- NEW: Background is executed N times, now part of scenarios.
-        duration = 0.0
+        feature_duration = 0.0
         for scenario in self.scenarios:
-            duration += scenario.duration
-        return duration
+            feature_duration += scenario.duration
+        return feature_duration
 
     def walk_scenarios(self, with_outlines=False):
         """
@@ -373,7 +390,7 @@ class Feature(TagStatement, Replayable):
         :param config:  Runner configuration to use (optional).
         :return: True, if scenario should run. False, otherwise.
         """
-        answer = self.status != "skipped"
+        answer = not self.should_skip
         if answer and config:
             answer = self.should_run_with_tags(config.tags)
         return answer
@@ -400,11 +417,17 @@ class Feature(TagStatement, Replayable):
         """
         Marks this feature (and all its scenarios and steps) as skipped.
         """
+        self._cached_status = None
+        self.should_skip = True
         for scenario in self.scenarios:
             scenario.mark_skipped()
+        else:
+            # -- SPECIAL CASE: Feature without scenarios
+            self._cached_status = "skipped"
         assert self.status == "skipped"
 
     def run(self, runner):
+        self._cached_status = None
         runner.context._push()
         runner.context.feature = self
 
@@ -427,7 +450,7 @@ class Feature(TagStatement, Replayable):
                 formatter.background(self.background)
 
         failed_count = 0
-        for scenario in self:
+        for scenario in self.scenarios:
             # -- OPTIONAL: Select scenario by name (regular expressions).
             if (runner.config.name and
                     not runner.config.name_re.search(scenario.name)):
@@ -440,6 +463,10 @@ class Feature(TagStatement, Replayable):
                 if runner.config.stop or runner.aborted:
                     # -- FAIL-EARLY: Stop after first failure.
                     break
+        else:
+            if not run_feature:
+                # -- SPECIAL CASE: Feature without scenarios:
+                self._cached_status = 'skipped'
 
         if run_feature:
             runner.run_hook('after_feature', runner.context, self)
@@ -492,7 +519,6 @@ class Background(BasicStatement, Replayable):
     '''
     type = "background"
 
-    # XXX-JE-ORIG: def __init__(self, filename, line, keyword, name, steps=[]):
     def __init__(self, filename, line, keyword, name, steps=None):
         super(Background, self).__init__(filename, line, keyword, name)
         self.steps = steps or []
@@ -511,7 +537,7 @@ class Background(BasicStatement, Replayable):
         return duration
 
 
-class Scenario(TagStatement, Replayable):
+class Scenario(TagAndStatusStatement, Replayable):
     '''A `scenario`_ parsed from a *feature file*.
 
     The attributes are:
@@ -577,43 +603,75 @@ class Scenario(TagStatement, Replayable):
     '''
     type = "scenario"
 
-    def __init__(self, filename, line, keyword, name, tags=[], steps=[],
+    def __init__(self, filename, line, keyword, name, tags=None, steps=None,
                  description=None):
+        tags = tags or []
         super(Scenario, self).__init__(filename, line, keyword, name, tags)
         self.description = description or []
         self.steps = steps or []
         self.background = None
         self.feature = None  # REFER-TO: owner=Feature
+        self._background_steps = None
         self._row = None
-        self.should_skip = None
+        self.was_dry_run = False
         self.stderr = None
         self.stdout = None
+
+    def reset(self):
+        '''
+        Reset the internal data to reintroduce new-born state just after the
+        ctor was called.
+        '''
+        super(Scenario, self).reset()
+        self._row = None
         self.was_dry_run = False
+        self.stderr = None
+        self.stdout = None
+        for step in self.all_steps:
+            step.reset()
+
+    @property
+    def background_steps(self):
+        '''
+        Provide background steps if feature has a background.
+        Lazy init that copies the background steps.
+
+        Note that a copy of the background steps is needed to ensure
+        that the background step status is specific to the scenario.
+
+        :return:  List of background steps or empty list
+        '''
+        if self._background_steps is None:
+            # -- LAZY-INIT (need copy of background.steps):
+            # Each scenario needs own background.steps status.
+            # Otherwise, background step status of the last scenario is used.
+            steps = []
+            if self.background:
+                steps = [copy.copy(step) for step in self.background.steps]
+            self._background_steps = steps
+        return self._background_steps
+
+    @property
+    def all_steps(self):
+        """Returns iterator to all steps, including background steps if any."""
+        if self.background is not None:
+            return itertools.chain(self.background_steps, self.steps)
+        else:
+            return iter(self.steps)
 
     def __repr__(self):
         return '<Scenario "%s">' % self.name
 
     def __iter__(self):
-        if self.background is not None:
-            return itertools.chain(self.background, self.steps)
-        else:
-            return iter(self.steps)
+        # XXX return iter(self.all_steps)
+        return self.all_steps
 
-    @property
-    def all_steps(self):
-        """Returns iterator to all steps, including background steps if any."""
-        return self.__iter__()
-
-    @property
-    def status(self):
-        if self.should_skip:
-            # -- PERFORMANCE SHORTCUT: Scenario(Outline) is marked as skipped.
-            return 'skipped'
-
-        for step in self.steps:
-            if step.status == 'failed':
-                return 'failed'
-            elif step.status == 'undefined':
+    def compute_status(self):
+        """Compute the status of the scenario from its steps.
+        :return: Computed status (as string).
+        """
+        for step in self.all_steps:
+            if step.status == 'undefined':
                 if self.was_dry_run:
                     # -- SPECIAL CASE: In dry-run with undefined-step discovery
                     #    Undefined steps should not cause failed scenario.
@@ -621,19 +679,24 @@ class Scenario(TagStatement, Replayable):
                 else:
                     # -- NORMALLY: Undefined steps cause failed scenario.
                     return 'failed'
-            elif step.status == 'skipped':
-                return 'skipped'
-            elif step.status == 'untested':
-                return 'untested'
+            elif step.status != 'passed':
+                assert step.status in ('failed', 'skipped', 'untested')
+                return step.status
+            #elif step.status == 'failed':
+            #    return 'failed'
+            #elif step.status == 'skipped':
+            #    return 'skipped'
+            #elif step.status == 'untested':
+            #    return 'untested'
         return 'passed'
 
     @property
     def duration(self):
         # -- ORIG: for step in self.steps:  Background steps were excluded.
-        duration = 0
+        scenario_duration = 0
         for step in self.all_steps:
-            duration += step.duration
-        return duration
+            scenario_duration += step.duration
+        return scenario_duration
 
     @property
     def effective_tags(self):
@@ -677,15 +740,20 @@ class Scenario(TagStatement, Replayable):
         """
         Marks this scenario (and all its steps) as skipped.
         """
+        self._cached_status = None
         self.should_skip = True
-        for step in self:
+        for step in self.all_steps:
             assert step.status == "untested" or step.status == "skipped"
             step.status = "skipped"
+        else:
+            # -- SPECIAL CASE: Scenario without steps
+            self._cached_status = "skipped"
         assert self.status == "skipped", "OOPS: scenario.status=%s" % self.status
 
     def run(self, runner):
         # pylint: disable=W0212
         #   W0212   Access to a protected member: runner.context._push()/._pop()
+        self._cached_status = None
         failed = False
         run_scenario = self.should_run(runner.config)
         run_steps = run_scenario and not runner.config.dry_run
@@ -721,6 +789,7 @@ class Scenario(TagStatement, Replayable):
                     run_steps = False
                     failed = True
                     runner.context._set_root_attribute('failed', True)
+                    self._cached_status = 'failed'
             elif failed or dry_run_scenario:
                 # -- SKIP STEPS: After failure/undefined-step occurred.
                 # BUT: Detect all remaining undefined steps.
@@ -730,11 +799,15 @@ class Scenario(TagStatement, Replayable):
                 found_step = step_registry.registry.find_match(step)
                 if not found_step:
                     step.status = 'undefined'
-                    runner.undefined.append(step)
+                    runner.undefined_steps.append(step)
             else:
                 # -- SKIP STEPS: For disabled scenario.
                 # NOTE: Undefined steps are not detected (by intention).
                 step.status = 'skipped'
+        else:
+            if not run_scenario:
+                # -- SPECIAL CASE: Scenario without steps.
+                self._cached_status = 'skipped'
 
         # Attach the stdout and stderr if generate Junit report
         if runner.config.junit:
@@ -825,13 +898,20 @@ class ScenarioOutline(Scenario):
     '''
     type = "scenario_outline"
 
-
-    def __init__(self, filename, line, keyword, name, tags=[],
-                 steps=[], examples=[], description=None):
+    def __init__(self, filename, line, keyword, name, tags=None,
+                 steps=None, examples=None, description=None):
         super(ScenarioOutline, self).__init__(filename, line, keyword, name,
                                               tags, steps, description)
         self.examples = examples or []
         self._scenarios = []
+
+    def reset(self):
+        '''
+        Reset runtime temporary data like before a test run.
+        '''
+        super(ScenarioOutline, self).reset()
+        for scenario in self.scenarios:
+            scenario.reset()
 
     @property
     def scenarios(self):
@@ -852,7 +932,6 @@ class ScenarioOutline(Scenario):
                 scenario.background = self.background
                 scenario._row = row
                 self._scenarios.append(scenario)
-
         return self._scenarios
 
     def __repr__(self):
@@ -861,33 +940,42 @@ class ScenarioOutline(Scenario):
     def __iter__(self):
         return iter(self.scenarios)
 
-    @property
-    def status(self):
+    def compute_status(self):
         for scenario in self.scenarios:
-            if scenario.status == 'failed':
-                return 'failed'
-            if scenario.status == 'skipped':
-                return 'skipped'
-            if scenario.status == 'untested':
-                return 'untested'
+            scenario_status = scenario.status
+            if scenario_status != 'passed':
+                assert scenario_status in ('failed', 'skipped', 'untested')
+                return scenario_status
+            #if scenario.status == 'failed':
+            #    return 'failed'
+            #elif scenario.status == 'skipped':
+            #    return 'skipped'
+            #elif scenario.status == 'untested':
+            #    return 'untested'
         return 'passed'
 
     @property
     def duration(self):
-        duration = 0
+        outline_duration = 0
         for scenario in self.scenarios:
-            duration += scenario.duration
-        return duration
+            outline_duration += scenario.duration
+        return outline_duration
 
     def mark_skipped(self):
         """
         Marks this scenario outline (and all its scenarios/steps) as skipped.
         """
+        self._cached_status = None
+        self.should_skip = True
         for scenario in self.scenarios:
             scenario.mark_skipped()
+        else:
+            # -- SPECIAL CASE: ScenarioOutline without scenarios/examples
+            self._cached_status = "skipped"
         assert self.status == "skipped"
 
     def run(self, runner):
+        self._cached_status = None
         failed_count = 0
         for scenario in self.scenarios:
             runner.context._set_root_attribute('active_outline', scenario._row)
@@ -897,7 +985,6 @@ class ScenarioOutline(Scenario):
                 if runner.config.stop or runner.aborted:
                     # -- FAIL-EARLY: Stop after first failure.
                     break
-
         runner.context._set_root_attribute('active_outline', None)
         return failed_count > 0
 
@@ -1017,6 +1104,13 @@ class Step(BasicStatement, Replayable):
         self.error_message = None
         self.exception = None
 
+    def reset(self):
+        '''Reset temporary runtime data to reach clean state again.'''
+        self.status = 'untested'
+        self.duration = 0.0
+        self.error_message = None
+        self.exception = None
+
     def __repr__(self):
         return '<%s "%s">' % (self.step_type, self.name)
 
@@ -1046,7 +1140,7 @@ class Step(BasicStatement, Replayable):
         # access module var here to allow test mocking to work
         match = step_registry.registry.find_match(self)
         if match is None:
-            runner.undefined.append(self)
+            runner.undefined_steps.append(self)
             if not quiet:
                 for formatter in runner.formatters:
                     formatter.match(NoMatch())
@@ -1065,7 +1159,8 @@ class Step(BasicStatement, Replayable):
                 formatter.match(match)
 
         runner.run_hook('before_step', runner.context, self)
-        runner.start_capture()
+        if capture:
+            runner.start_capture()
 
         try:
             start = time.time()
@@ -1095,8 +1190,8 @@ class Step(BasicStatement, Replayable):
             self.exception = e
 
         self.duration = time.time() - start
-
-        runner.stop_capture()
+        if capture:
+            runner.stop_capture()
 
         # flesh out the failure with details
         if self.status == 'failed':
@@ -1159,15 +1254,14 @@ class Table(Replayable):
     #   R0921   Abstract class is not referenced.
     type = "table"
 
-    def __init__(self, headings, line=None, rows=[]):
+    def __init__(self, headings, line=None, rows=None):
         Replayable.__init__(self)
         self.headings = headings
         self.line = line
         self.rows = []
-        if rows is None:
-            rows = []
-        for row in rows:
-            self.add_row(row, line)
+        if rows:
+            for row in rows:
+                self.add_row(row, line)
 
     def add_row(self, row, line=None):
         self.rows.append(Row(self.headings, row, line))
@@ -1529,3 +1623,13 @@ class NoMatch(Match):
         self.func = None
         self.arguments = []
         self.location = None
+
+
+def reset_model(model_elements):
+    """
+    Reset the test run information stored in model elements.
+
+    :param model_elements:  List of model elements (Feature, Scenario, ...)
+    """
+    for model_element in model_elements:
+        model_element.reset()