+++ test/test_log_capture.py
@@ -10,6 +10,7 @@ from __future__ import with_statement
 
 from nose.tools import *
 from mock import patch
+
 from behave.log_capture import LoggingCapture
 import unittest