+++ CHANGES.rst
@@ -21,15 +21,18 @@ IMPROVEMENT:
 
   * issue #108: behave.main() can be called with command-line args (provided by: medwards, jenisys)
   * issue #172: Subfolders in junit XML filenames (provided by: roignac).
+  * Simple test runner to run behave tests from "setup.py"
 
 FIXED:
 
   * issue #143: Logging starts with a StreamHandler way too early (provided by: jtatum, jenisys).
+  * issue #175: Scenario isn't marked as 'failed' when Background step fails
   * issue #177: Cannot setup logging_format
   * issue #184: TypeError when running behave with --include option (provided by: s1ider).
   * issue #186: ScenarioOutline uses wrong return value when if fails (provided by: mdavezac)
   * issue #188: Better diagnostics if nested step is undefined
   * issue #191: Using context.execute_steps() may change context.table/.text
+  * issue #194: Nested steps prevent that original stdout/stderr is restored
 
 
 Version: 1.2.3 (2013-07-08)